requirements.txt
pip
requests
openai
google-genai
types